requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-functions
azure-core
requests
azure-storage-blob
fitz
openai
openai-batch
dotenv
fastapi
uvicorn
python-multipart
pyodbc
azure-ai-formrecognizer